+++ src/main/java/data/mapper/ReviewMapper.java
@@ -1,5 +1,6 @@
 package data.mapper;
 
+import data.dto.ReviewDto;
 import data.dto.TagDto;
 import org.apache.ibatis.annotations.Mapper;
 
@@ -8,4 +9,6 @@ import java.util.List;
 @Mapper
 public interface ReviewMapper {
     int selectReviewCount(int num);
+    List<ReviewDto> getReviewList();
+
 }
