+++ src/main/resources/mapper/TagSql.xml
@@ -6,4 +6,5 @@
         select tname from tag where roomNum=#{num}
     </select>
 
+
 </mapper>
